Fix final compilation errors
- Fixed rows variable shadowing in slow_query_analyzer.go
- Fixed type mismatch in unified_database_monitor.go (int64 vs float64)

Note: Float truncation warnings in performance_report_generator.go
are likely false positives - the types are correct (time.Duration).
These may need investigation but don't block compilation.

+++ internal/database/unified_database_monitor.go
@@ -386,7 +386,7 @@ func (udm *UnifiedDatabaseMonitor) calculateQueryStats(metrics *DatabaseMetrics)
 
 	// Find max query time
 	for _, stats := range udm.queryStats {
-		if stats.MaxTime.Milliseconds() > metrics.MaxQueryTime {
+		if float64(stats.MaxTime.Milliseconds()) > metrics.MaxQueryTime {
 			metrics.MaxQueryTime = float64(stats.MaxTime.Milliseconds())
 		}
 	}
